+++ src/components/Editor/Editor.tsx
@@ -4,13 +4,13 @@ import {
 } from "./DocumentSelector/DocumentSelector";
 import "./Editor.scss";
 import { useEditorConfig } from "./config";
-import { DevComponentTestPlugin } from "./plugins/Dev/DevComponentTestPlugin";
-import { DevToolbarPlugin } from "./plugins/Dev/DevToolbarPlugin";
-import { IndentationPlugin } from "./plugins/IndentationPlugin";
-import { NoteControlsPlugin } from "./plugins/NoteControlsPlugin";
-import { NotesPlugin } from "./plugins/NotesPlugin/NotesPlugin";
-import { QuickMenuPlugin } from "./plugins/QuickMenuPlugin";
-import { RemdoAutoLinkPlugin } from "./plugins/RemdoAutoLinkPlugin";
+import { DevComponentTestPlugin } from "./plugins/dev/DevComponentTestPlugin";
+import { DevToolbarPlugin } from "./plugins/dev/DevToolbarPlugin";
+import { IndentationPlugin } from "./plugins/remdo/IndentationPlugin";
+import { NoteControlsPlugin } from "./plugins/remdo/NoteControlsPlugin";
+import { NotesPlugin } from "./plugins/RemdoPlugin";
+import { QuickMenuPlugin } from "./plugins/remdo/QuickMenuPlugin";
+import { RemdoAutoLinkPlugin } from "./plugins/remdo/RemdoAutoLinkPlugin";
 import { ClearEditorPlugin } from "@lexical/react/LexicalClearEditorPlugin";
 import LexicalClickableLinkPlugin from "@lexical/react/LexicalClickableLinkPlugin";
 import { CollaborationPlugin } from "@lexical/react/LexicalCollaborationPlugin";
