+++ src/components/Editor/plugins/remdo/utils/patches.ts
@@ -1,4 +1,4 @@
-import { NotesState, Note } from "@/components/Editor/api";
+import { NotesState, Note } from "@/components/Editor/plugins/remdo/utils/api";
 import { patch } from "@/utils";
 
 export function applyNodePatches(NodeType: any) {
